I'm struggling here
I've got a few cards that are failing UNIQUE name... news to me, but there is
at least 2 cards called "Bind". I don't know why, but that annoys be a lot!

This is a crap mid - development commit where I'm not sure how best
tot handle cards with 2 "faces" (double sided, or adventures at least)
in the easiest whay I can manage it.